live/requirement.txt
# Please run pip freeze to findout the version of each newly introduced package

Django==1.10.6
django-chartit==0.2.9
django-common-helpers==0.9.0
django-cron==0.5.0
django-redis==4.7.0
django-filter==1.0.2
djangorestframework==3.6.2
django-dynamic-preferences==1.1
leancloud-sdk==1.10.0
#mysqlclient==1.3.10
oss2==2.2.3
Pillow==4.0.0
pypinyin==0.18.1
redis==2.10.5
requests==2.13.0
WeRoBot==1.1.1
msgpack-python==0.4.8
celery==4.0.2
raven==6.0.0
django-cors-headers==2.0.2
requests-toolbelt==0.7.1
django-rest-elasticsearch==0.2.0
rsa==3.4.2
oss2==2.2.3
celery==4.0.2